@@ -19,17 +19,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.8",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.2.8",
+    "@angular/compiler": "^18.2.8",
+    "@angular/core": "^18.2.8",
+    "@angular/forms": "^18.2.8",
     "@angular/material": "^17.3.10",
     "@angular/material-moment-adapter": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.8",
+    "@angular/platform-browser-dynamic": "^18.2.8",
+    "@angular/router": "^18.2.8",
     "@codeffekt/ce-barcode": "^16.0.0",
     "@codeffekt/ce-code-editor": "^16.0.0",
     "@codeffekt/ce-core-data": "^1.2.0",
@@ -55,10 +55,10 @@
     "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.10",
-    "@angular-devkit/schematics": "^17.3.10",
-    "@angular/cli": "^17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.8",
+    "@angular-devkit/schematics": "^18.2.8",
+    "@angular/cli": "^18.2.8",
+    "@angular/compiler-cli": "^18.2.8",
     "@types/jasmine": "^3.6.3",
     "@types/node": "^12.19.14",
     "eslint": "^8.13.0",
@@ -69,7 +69,7 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.2.1",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "typescript": "^5.4.5"
