@@ -1,21 +1,14 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { AngularSvgIconModule, SvgIconRegistryService } from 'angular-svg-icon';
 import { CeIllustrationComponent } from './illustration.component';
 
-@NgModule({
-    imports: [
-        CommonModule,
-        AngularSvgIconModule.forRoot(),
-        HttpClientModule,
-    ],
-    exports: [
+@NgModule({ exports: [
         CeIllustrationComponent
     ],
     declarations: [
         CeIllustrationComponent
-    ],
-    providers: [],
-})
+    ], imports: [CommonModule,
+        AngularSvgIconModule.forRoot()], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class CeIllustrationModule { }
